11 connect server to client

+++ client/src/components/ui/popoverMenu.tsx
@@ -36,11 +36,24 @@ function PopoverMenu({ type, data, children }: PopoverMenuProps) {
 	const [open, setOpen] = useState(false);
 	const [isEditName, setIsEditName] = useState(false);
 
-	function handleClickOption(option: string) {
+	async function handleClickOption(option: string) {
 		if (type === "task") {
 			switch (option) {
 				case "delete":
-					deleteTask(data.id);
+					try {
+						const response = await fetch(
+							`http://localhost:3000/api/v1/tasks/${data.id}`,
+							{
+								method: "DELETE",
+							},
+						);
+
+						if (!response.ok) throw new Error("Failed to delete the task");
+
+						deleteTask(data.id);
+					} catch (error) {
+						console.error("Error deleting the task", error);
+					}
 					break;
 				case "edit":
 					setOpen(true);
@@ -49,10 +62,22 @@ function PopoverMenu({ type, data, children }: PopoverMenuProps) {
 		} else if (type === "list") {
 			switch (option) {
 				case "delete":
-					deleteGroup(data.id);
-					setTasks((prevTasks) =>
-						prevTasks.filter((task) => task.group_id !== data.id),
-					);
+					try {
+						const response = await fetch(
+							`http://localhost:3000/api/v1/groups/${data.id}`,
+							{ method: "DELETE" },
+						);
+
+						if (!response.ok) throw new Error("Failed to delete the group");
+
+						deleteGroup(data.id);
+						setTasks((prevTasks) =>
+							prevTasks.filter((task) => task.groupId !== data.id),
+						);
+					} catch (error) {
+						console.error("Error deleting the group", error);
+					}
+
 					break;
 				case "edit":
 					setIsEditName(true);
@@ -61,10 +86,23 @@ function PopoverMenu({ type, data, children }: PopoverMenuProps) {
 		} else if (type === "column") {
 			switch (option) {
 				case "delete":
-					deleteStatus(data.id);
-					setTasks((prevTasks) =>
-						prevTasks.filter((task) => task.status_id !== data.id),
-					);
+					try {
+						const response = await fetch(
+							`http://localhost:3000/api/v1/statuses/${data.id}`,
+							{
+								method: "DELETE",
+							},
+						);
+
+						if (!response.ok) throw new Error("Failed to delete the status");
+
+						deleteStatus(data.id);
+						setTasks((prevTasks) =>
+							prevTasks.filter((task) => task.statusId !== data.id),
+						);
+					} catch (error) {
+						console.error("Error deleting the group", error);
+					}
 					break;
 				case "edit":
 					setIsEditName(true);
